another bf

+++ AFQ/api.py
@@ -1170,7 +1170,7 @@ class AFQ(object):
             '_seed_mask.nii.gz')
         if not op.exists(seed_file):
             self.log_and_save_nii(
-                nib.Nifti1Image(seed_mask, row["dwi_affine"]),
+                nib.Nifti1Image(seed_mask.astype(np.float32), row["dwi_affine"]),
                 seed_file)
             afd.write_json(self._get_fname(row, '_seed_mask.json'), seed_mask_desc)
         return seed_file
@@ -1187,7 +1187,7 @@ class AFQ(object):
             '_stop_mask.nii.gz')
         if not op.exists(stop_file):
             self.log_and_save_nii(
-                nib.Nifti1Image(stop_mask, row["dwi_affine"]),
+                nib.Nifti1Image(stop_mask.astype(np.float32), row["dwi_affine"]),
                 stop_file)
             afd.write_json(self._get_fname(row, '_stop_mask.json'), stop_mask_desc)
         return stop_file
